LayoutTests/platform/gtk/accessibility/svg-image-expected.txt

 
This tests that SVG images are accessible elements and they have the same attributes as real images.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgImage.role is realImage.role
PASS svgImage.title is realImage.title
SVG Image Role: AXRole: AXImage
SVG Image Title: AXTitle: TestImage
PASS successfullyParsed is true

TEST COMPLETE
